Cleanup pkla code and support logind
In Trusty, the policies for suspend and hibernate have moved from
upower to logind.  So we add groups to the pkla files to disable
those logind actions.  Note that logind differentiates between
actions for seats that support single sessions vs multi sessions,
and all Unity sessions are multi-user sessions, by definition.
(Because Unity.)

Also, merge the pkla files to a single file in /var/lib/polkit-1,
with a hierarchical name, and add maintscript-helper code in the
maintscripts to punt the old etc conffiles on upgrade.